Rename PreprocessorListener to DefaultPreprocessorListener, and add PreprocessorListener interface
